requirements.txt
aiofiles==0.8.0
asyncpg==0.25.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
Flask==2.0.2
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.2
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.27
transliterate==1.10.2
urllib3==1.26.8
Werkzeug==2.0.2
